requirements.txt
Django==4.2.18
pysentimiento
python-decouple
django-htmx

